iox-#1032 Test error verification
